requirements.txt
python-telegram-bot>=20.7
aiohttp>=3.9.1
aiosqlite>=0.19.0
SQLAlchemy>=2.0.25
alembic>=1.13.1
greenlet>=3.0.3
APScheduler>=3.10.4
pytz>=2023.3
python-dotenv>=1.0.0
pillow>=10.1.0
tenacity>=8.2.0
asyncpg>=0.29.0